TestProject/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
bleach==6.0.0
bs4==0.0.1
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.0.1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
Django==4.1.7
djangorestframework==3.14.0
drf-api-logger==1.1.12
et-xmlfile==1.1.0
filelock==3.7.0
identify==2.5.1
idna==3.4
MarkupPy==1.14
nodeenv==1.6.0
numpy==1.24.1
odfpy==1.4.1
openpyxl==3.1.0
pandas==1.5.3
platformdirs==2.5.2
pre-commit==2.19.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tablib==3.3.0
toml==0.10.2
urllib3==1.26.14
virtualenv==20.14.1
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
